use enum and move all to constants
