+++ packages/providers/google/src/constants.ts
@@ -4,7 +4,30 @@ import {
   type SafetySetting,
 } from "@google/genai";
 import type { MessageType } from "@llamaindex/core/llms";
-import { GEMINI_MESSAGE_ROLE, GEMINI_MODEL } from "./types.js";
+
+export enum GEMINI_MODEL {
+  GEMINI_PRO = "gemini-pro",
+  GEMINI_PRO_VISION = "gemini-pro-vision",
+  GEMINI_PRO_LATEST = "gemini-1.5-pro-latest",
+  GEMINI_PRO_FLASH_LATEST = "gemini-1.5-flash-latest",
+  GEMINI_PRO_1_5_PRO_PREVIEW = "gemini-1.5-pro-preview-0514",
+  GEMINI_PRO_1_5_FLASH_PREVIEW = "gemini-1.5-flash-preview-0514",
+  GEMINI_PRO_1_5 = "gemini-1.5-pro-001",
+  GEMINI_PRO_1_5_FLASH = "gemini-1.5-flash-001",
+  // Note: should be switched to -latest suffix when google supports it
+  GEMINI_PRO_1_5_LATEST = "gemini-1.5-pro-002",
+  GEMINI_PRO_1_5_FLASH_LATEST = "gemini-1.5-flash-002",
+  GEMINI_2_0_FLASH_EXPERIMENTAL = "gemini-2.0-flash-exp",
+  GEMINI_2_0_FLASH = "gemini-2.0-flash-001",
+  GEMINI_2_0_FLASH_LITE = "gemini-2.0-flash-lite-001",
+  GEMINI_2_0_FLASH_LITE_PREVIEW = "gemini-2.0-flash-lite-preview-02-05",
+  GEMINI_2_0_FLASH_THINKING_EXP = "gemini-2.0-flash-thinking-exp-01-21",
+  GEMINI_2_0_PRO_EXPERIMENTAL = "gemini-2.0-pro-exp-02-05",
+  GEMINI_2_0_FLASH_LIVE = "gemini-2.0-flash-live-001",
+  GEMINI_2_5_PRO_PREVIEW = "gemini-2.5-pro-preview-03-25",
+  GEMINI_2_5_PRO_PREVIEW_LATEST = "gemini-2.5-pro-preview-06-05",
+  GEMINI_2_5_FLASH_PREVIEW = "gemini-2.5-flash-preview-05-20",
+}
 
 export const GEMINI_MODEL_INFO_MAP: Record<
   GEMINI_MODEL,
@@ -81,6 +104,23 @@ export const DEFAULT_SAFETY_SETTINGS: SafetySetting[] = [
     threshold: HarmBlockThreshold.BLOCK_NONE,
   },
 ];
+
+export enum GEMINI_MESSAGE_ROLE {
+  USER = "user",
+  MODEL = "model",
+  FUNCTION = "function",
+}
+
+export enum GEMINI_VOICE_NAME {
+  PUCK = "Puck",
+  CHARON = "Charon",
+  FENRIR = "Fenrir",
+  AOEDE = "Aoede",
+  LEDA = "Leda",
+  KORE = "Kore",
+  ORUS = "Orus",
+  ZEPHYR = "Zephyr",
+}
 
 // Gemini only has user and model roles. Put the rest in user role.
 export const ROLES_TO_GEMINI: Record<MessageType, GEMINI_MESSAGE_ROLE> = {
