+++ packages/providers/google/src/index.ts
@@ -1,12 +1,12 @@
 export * from "./base";
 
+export * from "./constants";
 export * from "./live";
 export {
   GoogleStudio,
   Modality,
   getGoogleStudioInlineData,
 } from "./studio/index.js";
-export * from "./types";
 export * from "./utils";
 
 export * from "./GeminiEmbedding";
